New project file structure.
