+++ .gitignore
@@ -2,4 +2,3 @@ node_modules
 package-lock.json
 npm-debug.log
 .DS_Store
-lib
