+++ lib/constants.js
@@ -1,0 +1,12 @@
+module.exports = {
+  DEFAULTS: {
+    sourceGlob: '**/*.{mjs,js}',
+    markdownPath: 'readme.md',
+    targetHeading: 'API'
+  },
+  MEMBERSHIPS: {
+    '.': 'static',
+    '#': 'instance',
+    '~': 'inner'
+  }
+}
